settings from __init__.py to app.py
